Fixed bug where views would vibrate slightly during a resize on Apple
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/renaissance/trunk@26328 72102866-910b-0410-8b05-ffd578937521

+++ Source/AutoLayout/GSAutoLayoutProportionalManager.m
@@ -116,7 +116,7 @@
 
   /* Now, compute the _minimumLayout of all line parts.  */
   {
-    int position = 0;
+    float position = 0;
     int i, count = [_lineParts count];
     
     for (i = 0; i < count; i++)
@@ -197,7 +197,7 @@
 
   /* Now, compute the _layout of all line parts.  */
   {
-    int position = 0;
+    float position = 0;
     int i, count = [_lineParts count];
     
     for (i = 0; i < count; i++)
